requirements.txt
alembic==1.14.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
certifi==2026.1.4
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.4
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==2.1.5
packaging==26.0
pkg_resources==0.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pytest==8.3.5
python-dotenv==1.0.1
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.44.0
tomli==2.4.0
typing_extensions==4.13.2
uvicorn==0.33.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
